+++ Jellyfin.Plugin.OpenSubtitles/Jellyfin.Plugin.OpenSubtitles.csproj
@@ -12,7 +12,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="../RESTOpenSubtitlesHandler/RESTOpenSubtitlesHandler.csproj" />
+    <ProjectReference Include="../OpenSubtitlesHandler/OpenSubtitlesHandler.csproj" />
   </ItemGroup>
 
   <ItemGroup>
