+++ OpenSubtitlesHandler/Util.cs
@@ -10,7 +10,7 @@ using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace RESTOpenSubtitlesHandler {
+namespace OpenSubtitlesHandler {
     public static class Util {
         private static HttpClient HttpClient = new HttpClient();
         public static Action<string> OnHTTPUpdate = _ => {};
